Fix mpio_pmem_pool_size type to use size_t instead of int

+++ src/mpi/romio/adio/include/adioi.h
@@ -53,7 +53,7 @@ struct ADIOI_Hints_struct {
       int io_nthreads;
       int pool_per_rank;
       char* pool_list;
-      int pool_size;
+      size_t pool_size;
     } pmem;
     union {
         struct {
